move createHelixPropertyStore method to ambry-api
