+++ ambry-clustermap/src/test/java/com.github.ambry.clustermap/ClusterMapUtilsTest.java
@@ -13,82 +13,21 @@
  */
 package com.github.ambry.clustermap;
 
-import com.github.ambry.config.HelixPropertyStoreConfig;
-import com.github.ambry.config.VerifiableProperties;
 import com.github.ambry.network.Port;
 import com.github.ambry.network.PortType;
 import com.github.ambry.utils.UtilsTest;
-import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Properties;
 import java.util.Set;
-import org.apache.helix.AccessOption;
-import org.apache.helix.ZNRecord;
-import org.apache.helix.store.HelixPropertyStore;
 import org.junit.Test;
 
-import static com.github.ambry.utils.TestUtils.*;
 import static org.junit.Assert.*;
 
 
 public class ClusterMapUtilsTest {
-
-  /**
-   * Tests for {@link ClusterMapUtils#createHelixPropertyStore(String, HelixPropertyStoreConfig, List)}.
-   */
-  @Test
-  public void createHelixPropertyStoreTest() throws IOException {
-    Properties storeProps = new Properties();
-    storeProps.setProperty("helix.property.store.root.path", "/Ambry-Test");
-    HelixPropertyStoreConfig propertyStoreConfig = new HelixPropertyStoreConfig(new VerifiableProperties(storeProps));
-    String tempDirPath = getTempDir("clusterMapUtils-");
-    ZkInfo zkInfo = new ZkInfo(tempDirPath, "DC1", (byte) 0, 2200, true);
-
-    try {
-      ClusterMapUtils.createHelixPropertyStore(null, propertyStoreConfig, Collections.EMPTY_LIST);
-      fail("create HelixPropertyStore with invalid arguments should fail");
-    } catch (IllegalArgumentException e) {
-      //expected
-    }
-
-    try {
-      ClusterMapUtils.createHelixPropertyStore("", propertyStoreConfig, Collections.EMPTY_LIST);
-      fail("create HelixPropertyStore with invalid arguments should fail");
-    } catch (IllegalArgumentException e) {
-      //expected
-    }
-
-    try {
-      ClusterMapUtils.createHelixPropertyStore("localhost:" + zkInfo.getPort(), null, Collections.EMPTY_LIST);
-      fail("create HelixPropertyStore with invalid arguments should fail");
-    } catch (IllegalArgumentException e) {
-      //expected
-    }
-
-    HelixPropertyStore<ZNRecord> propertyStore =
-        ClusterMapUtils.createHelixPropertyStore("localhost:" + zkInfo.getPort(), propertyStoreConfig,
-            Collections.singletonList(propertyStoreConfig.rootPath));
-    assertNotNull(propertyStore);
-
-    // Ensure the HelixPropertyStore works correctly
-    List<String> list = Arrays.asList("first", "second", "third");
-    String path = propertyStoreConfig.rootPath + ClusterMapUtils.PROPERTYSTORE_ZNODE_PATH;
-    ZNRecord znRecord = new ZNRecord(ClusterMapUtils.ZNODE_NAME);
-    znRecord.setListField("AmbryList", list);
-    if (!propertyStore.set(path, znRecord, AccessOption.PERSISTENT)) {
-      fail("Failed to set HelixPropertyStore");
-    }
-    // Verify path exists
-    assertTrue("The record path doesn't exist", propertyStore.exists(path, AccessOption.PERSISTENT));
-    // Verify record
-    ZNRecord result = propertyStore.get(path, null, AccessOption.PERSISTENT);
-    assertEquals("Mismatch in list content", new HashSet<>(list), new HashSet<>(result.getListField("AmbryList")));
-    zkInfo.shutdown();
-  }
 
   /**
    * Tests for {@link ClusterMapUtils#areAllReplicasForPartitionUp(PartitionId)}.
